// webview-ui/src/thor/model/Ptg.ts
export type Ptg  = {
    /**
     * 
     * @type {string}
     * @memberof Ptg
     */
    sheetId?: string;
    /**
     * 
     * @type {string}
     * @memberof Ptg
     */
    formulaId?: string;
    /**
     * 
     * @type {string}
     * @memberof Ptg
     */
    stringValue?: string;
    /**
     * 
     * @type {number}
     * @memberof Ptg
     */
    numberValue?: number;
    /**
     * 
     * @type {string}
     * @memberof Ptg
     */
    type?: PtgTypeEnum;
    /**
     * Unique identifier for object in the system
     * @type {string}
     * @memberof Ptg
     */
    id?: string;
    /**
     * UUID of owner of the object in the system
     * @type {string}
     * @memberof Ptg
     */
    ownerId?: string;
    /**
     * Date of object creation
     * @type {Date}
     * @memberof Ptg
     */
    createdDate?: Date;
    /**
     * Data, including hash of the key(s) used to encrypt this record.
     * @type {string}
     * @memberof Ptg
     */
    keyHash?: string;
    /**
     * Last user to access object
     * @type {string}
     * @memberof Ptg
     */
    lastAccessedById?: string;
    /**
     * Timestamp of last access of object
     * @type {Date}
     * @memberof Ptg
     */
    lastAccessedDate?: Date;
    /**
     * Unique identifier for user who last modifed the object in the system
     * @type {string}
     * @memberof Ptg
     */
    lastModifiedById?: string;
    /**
     * Date of last object modification
     * @type {Date}
     * @memberof Ptg
     */
    lastModifiedDate?: Date;
}
// ...
export enum PtgTypeEnum {
    PTGINT = 'ptgint',
    PTGSTR = 'ptgstr',
    PTGATTR = 'ptgattr',
    PTGFUNC = 'ptgfunc',
    PTGFUNCVAR = 'ptgfuncvar',
    PTGNAME = 'ptgname',
    PTGREF = 'ptgref',
    PTGAREA = 'ptgarea'
}